master, development troche poprawek do dodwania.aktualizacji i usuwania

+++ src/app/home/home.component.ts
@@ -3,7 +3,7 @@ import {Http, Response, Headers} from '@angular/http';
 import {CustomersService} from '../services/CustomersService/Customers.service';
 import {EmitterService} from '../services/EmitterService/Emitter.service';
 import {ICustomers} from '../models/customer';
-
+import {ActivatedRoute,Router} from '@angular/router';
 
 @Component({
   selector: 'app-customers',
@@ -12,9 +12,9 @@ import {ICustomers} from '../models/customer';
 })
 export class HomeComponent implements OnInit, OnChanges {
   customers: ICustomers[];
-  @Input() public listId: string;
+  @Input() public listId: any;
   @Input() editId: string;
-  constructor(private _customersService: CustomersService) {
+  constructor(private _customersService: CustomersService,private router: Router) {
 
   }
 
@@ -32,16 +32,21 @@ export class HomeComponent implements OnInit, OnChanges {
   }
 
   ngOnChanges(changes : any) {
+    console.log("ngOnChanges");
     EmitterService.get(this.listId).subscribe((customers: ICustomers[]) => {this.loadCustomers()});
   }
 
   deleteCustomer(id : string) {
+    var result = confirm("Are you sure to delete?")
+    if(result)
+    {
     this._customersService.removeCustomer(id).subscribe(
       customer => { EmitterService.get(id).emit(customer); 
       },
       err => {
         console.log(err);
       });
+      this.router.navigate(['/']);
+    }  
   }
-    
 }
